refactor: move components reaction settings in components package
