+++ src/main/scala/satify/view/utils/TextPaneUtils.scala
@@ -1,7 +1,7 @@
 package satify.view.utils
 
 import satify.dsl.Reflection.getDSLKeywords
-import satify.view.Constants.fontFamily
+import Constants.fontFamily
 
 import java.awt.{Color, Font}
 import javax.swing.text.{MutableAttributeSet, SimpleAttributeSet, StyleConstants}
